requirements.txt
torch
torchvision
gradio
numpy
opencv-python
Pillow
